separate benchmarks from functionality test
