+++ tests/gcn_test.py
@@ -1,7 +1,5 @@
-import pytest
 import torch
 from graph_nn import GCNLayer
-from main import run
 from torch_geometric.nn import GCNConv
 
 
@@ -26,24 +24,3 @@ def test_gcn_against_pyg(num_nodes: int, symmetric_adj_mat_idx: torch.Tensor):
     grad1 = pyg_gcn.lin.weight.grad
     grad2 = gcn.linear.weight.grad
     assert torch.allclose(grad1, grad2, atol=1e-5)
-
-
-@pytest.mark.parametrize("dataset", ["Cora", "CiteSeer"])
-def test_gcn_cora_citeseer(dataset: str):
-    test_acc = run(
-        model="gcn",
-        dataset=dataset,
-        num_epochs=200,
-        optimizer="Adam",
-        lr=1e-2,
-        weight_decay=0,
-        momentum=0,
-        l2_regularization=5e-4,
-        device="cpu",
-        hidden_dim=16,
-        dropout=0.5,
-    )
-    if dataset == "Cora":
-        assert test_acc > 0.81
-    elif dataset == "CiteSeer":
-        assert test_acc > 0.70
